LayoutTests/java/lc3/ConvertNumber/number-005-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion Preferred argument conversion:  undefined

FAIL TEST_CLASS.ambiguous(1) should be INT. Was BYTE.
PASS successfullyParsed is true

TEST COMPLETE
 
